Currency IDs for Chracter / Currency tooltips
This is patch 2 of a few.

This one shows the currency IDs in the window when moving the mouse over the icons on the Character / Currency frame.

There might be a better way to do this and I'm still looking around for one.  I've got hooks listening for "SetCurrencyByID" and "SetCurrencyTokenByID" to see where or how they're used.

Currencies from vendors (to buy honor gear for example) do not show the currency ID.  Looking into those frames next.